refactor custom firebase hooks
