--- src/hooks/useFirebase.js
@@ -1,16 +1,0 @@
-import { useState, useEffect } from "react";
-
-const useFirebase = (productsRef) => {
-  const [data, setData] = useState([]);
-
-  useEffect(() => {
-    const unsub = productsRef.on("value", (snapshot) => {
-      setData(snapshot.val());
-    });
-
-    return () => unsub;
-  }, [productsRef]);
-  return { data };
-};
-
-export default useFirebase;
